# mwavuli-backend/infra/openapi.yaml
openapi: 3.1.0
info:
  title: mwavuli API
  version: "0.1.0"
  description: >
    Tree identification, mapping, and community backend. Public endpoints expose
    only fuzzy locations; exact coordinates require ownership/staff and are
    audited.
servers:
  - url: https://api.mwavuli.app
  - url: http://localhost:8080
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
    LatLng:
      type: object
      properties: { lat: { type: number }, lng: { type: number } }
    Tree:
      type: object
      properties:
        id: { type: string, format: uuid }
        contributor: { type: string }
        commonName: { type: string }
        scientificName: { type: string, nullable: true }
        heightM: { type: number, nullable: true }
        girthM: { type: number, nullable: true }
        ageEstimate: { type: string, nullable: true }
        health: { type: string, enum: [healthy, stressed, dead, unknown] }
        features: { type: array, items: { type: string } }
        confidence: { type: integer, nullable: true }
        verified: { type: boolean }
        visibility: { type: string, enum: [public, followers, private] }
        isFuzzy: { type: boolean }
        likeCount: { type: integer }
        commentCount: { type: integer }
        createdAt: { type: string, format: date-time }
        fuzzyLocation: { $ref: '#/components/schemas/LatLng', nullable: true }
...
paths:
...
  /v1/feed:
    get:
      summary: Public feed (optionally within a map bbox)
      security: []
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 20 } }
        - { name: before, in: query, schema: { type: string, format: date-time } }
        - { name: bbox, in: query, description: "minLng,minLat,maxLng,maxLat", schema: { type: string } }
      responses:
        '200':
          description: items
          content:
            application/json:
              schema: { type: object, properties: { items: { type: array, items: { $ref: '#/components/schemas/Tree' } } } }
...
  /v1/trees/{id}/exact-location:
    get:
      summary: Exact coordinates — owner/staff only, AUDITED
      responses:
        '200': { description: lat/lng/accuracy }
        '403': { description: not permitted }
...